OSS player: handle changing end position correctly
If the slider is moved to the playbar while the player is running,
the playback needs to be stopped.  Handle this situation properly.

Additionally, select a low-latency buffer policy if possible (OSS4
only).  This will make it possible to move the slide near the playbar
without stopping playback (the current writing position is always in
advance of the playback position by the buffer size!)

Originally committed to SVN as r3498.