Fix some things in the po files
In cases where an English-specific change was made to a string (such as
fixing a misspelling or colour -> color), update the corresponding msgid
in the po.

Apply some of the punctuation changes where it's safe to. Many strings
lost some punctuation at the end, so simply update the msgid when the
translation didn't have any punctuation at the end, and if the
punctuation was the same, update both the msgid and msgstr (and leave it
up to the translator if the punctuation was different).

Update msgid and msgstr in the cases where the printf formatting
character was the only thing that changed.

This decreases the number of new strings by up to ~150.

Originally committed to SVN as r6686.