TimeEdit: handle Shift-Insert the same way as Ctrl-V
Shift-Insert pasted the text anyway, but without anyone "noticing",
which meant the TimeEdit's internal variable "time" didn't get updated,
causing TimeEdit::GetTime() to return the old time. Because of this,
pasting a time in the Jump To dialog with Shift-Insert didn't update the
frame number.

Originally committed to SVN as r6882.