Add syntax highlighting for drawings and vector clips
The highlighting distinguishes drawing commands from coordinates, and
colors x and y coordinates in different colors to make coordinates
easier to visually parse. Furthermore, in cubic Bezier curves, it
underlines the coordinates which corresponds to endpoints of the curves.

The highlighting colors are chosen to match the conventions of 3D
modeling/engine software (e.g. Blender, Unity, etc), i.e. red for X and
green for Y. This has the risk of being harder to tell apart for
colorblind people, but all colors can be configured.