Use parent's font for warning text in preferences
The original code used wxFONTFAMILY_SWISS to set a bold sans-serif
font for the warning message in Preferences/Advanced, assuming that:
1. SWISS always returns a sans-serif font.
2. The system UI font is universally sans-serif across all locales.

However, the first assumption is incorrect, and the second one *was*
incorrect. On modern Windows under Chinese locales, SWISS will return
SimSun or PMingLiU, which are serif fonts with embedded bitmaps that
are distributed with legacy Windows as the default interface fonts.

The commit fixes the issue by using the parent's font and making
it bold instead of directly constructing a wxFont with SWISS.