use system gtest if available
Package maintainers won't need to patch Aegisub to use the system
version. This also helps Aegisub compile on platforms where the
unmodified gtest source doesn't compile but a package is available, such
as OpenBSD.

I've tested the conditional branch where gtest is present in the system
on OpenBSD. I can't properly test the branch where it isn't present
because the subproject doesn't compile on OpenBSD, but I've checked that
it tries to compile the subproject.