Get closer to the Lua 5.2 C API
Aegisub requires Luajit implementing Lua 5.2, but the code was making a
mixture of Lua 5.1 and Lua 5.2 C API calls which happened to work
because luajit's standards compliance is loose. A call to `lua_objlen()`
remains because Luajit doesn't implement Lua 5.2 fully and is missing
`lua_rawlen()`.

Tested with some out of tree patches to make Aegisub work on an
unsupported platform (OpenBSD, official Lua).