Round instead of truncating in float audio converter
This is mainly to make the tests {float,double}_conversion more robust,
previously they could fail depending on the target CPU type (e.g. with
gcc -m32 -march=i386).
In principle this is now inconsistent with the other converters which
all still truncate, but this doesn't really make a difference in
practice, and making them round would add additional branching and (if
implemented naively) break some other tests, so this is probably fine.

Fixes #274.