Deduplicate default_{config,hotkey,menu}.json
Previously, default_config.json and osx/default_config.json were 95%
identical, and default_hotkey/default_menu weren't much better. This
ran the risk of adding some config option (/hotkey/menu entry) to the
default config but forgetting to add it to osx.

So, replace this with a system involving one base config file and one or
more overrides files (one per operating system and one for per-build
configuration). While the config system already supports loading
incremental overrides, the menu and hotkey systems don't, so it seems
simpler to do this with a simple python script at build time.

That said, the meson side of things did need a lot of
path/custom_target/etc juggling to get everything to work correctly and
with the proper regen dependencies. Once newer meson versions can be
required this can be simplified a bit.

For default_menu.json we don't define the menus that are overwritten for
osx in the base config file and instead make a default_menu_platform
file for all other operating systems. This reduces the risk of adding a
menu entry to one of these menus and forgetting to add it to osx.

The new diff files for osx were generated using
`combine-config.py --diff`.