Fix squished video on large video zoom
glViewport silently clamps viewport sizes to its internal caps, which
seem to be 16384 on most machines. If video zoom is implemented by
enlarging the viewport, this results in squished video on large video
zooms (which also only squishes the video and not the visual tools,
resulting in seemingly offset visual tools).

Fix this by applying the video zoom and pan via the projection matrix
instead and keeping the glViewport equal to the client size.