Allow configuring the actions of ctrl/shift/normal scroll on the video
Allow mapping scroll, ctrl+scroll, and shift+scroll each to resizing,
zooming, or panning the video, possibly with swapped axes.
Have the default settings keep normal scroll as a video display resize
to match past behavior (but this could be changed in the future - now
that video zooming is possible, users will likely want to resize the
window less often) but map ctrl+scroll to a zoom and shift+scroll to a
pan.

The panning action is mainly geared towards touchpads, where
middle-click+dragging is not possible. In other programs (e.g.
browsers), normal scroll being a vertical scroll (i.e. a pan) and
shift+scroll being a horizontal scroll seems to be the established
behavior - this can be configured by mapping scroll to a pan and
shift+scroll to a pan with swapped axes.

Previous iterations of this commit also had options to flip the X and/or
Y axes of pans. However, the current code (X being hardcoded as flipped,
Y being hardcoded as not flipped) matches vertical and horizontal
scrolling in other applications like browsers on all tested
configurations, both with natural scrolling on and off. So, to prevent
option bloat, these options are kept out for now. If users report the
pan directions not matching other applications on their setups, they
can always be added in again.