Allow translating choices for string options in preferences
6f7e314b6765ea51afae09062277baf7b5a5fcd4 made the FFMS2 log levels
translatable. Translating dropdown choices in the preferences dialog
worked fine for int options, but not for string options.
To keep these strings translatable but not break compatibility by
turning the option into an int option, make the OptionChoice function
able to translate string options itself, while storing the untranslated
values in the options.

This is quite a lot of added code just to make one niche option
translatable, but it allows turning more options into string options in
the future, which behave better when better when choices are added.
(On the other hand, int options behave better when choices are renamed,
but either way the upshot is that choices for string options can now be
translated.)